# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studify API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  streak: number;
  totalHours: number;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface Group {
  id: number;
  name: string;
  adminId: number;
  adminName: string;
  memberCount: number;
  createdAt: string;
}

export interface GroupMember {
  userId: number;
  name: string;
  streak: number;
  totalHours: number;
  isAdmin: boolean;
}

export interface CreateGroupBody {
  name: string;
}

export interface Post {
  id: number;
  groupId: number;
  userId: number;
  userName: string;
  subject: string;
  hours: number;
  /** @nullable */
  description: string | null;
  createdAt: string;
}

export interface CreatePostBody {
  subject: string;
  hours: number;
  /** @nullable */
  description?: string | null;
}

export interface RankingEntry {
  userId: number;
  name: string;
  weeklyHours: number;
  streak: number;
  rank: number;
}

export interface SendInviteBody {
  email: string;
}

export interface Invite {
  id: number;
  groupId: number;
  groupName: string;
  invitedById: number;
  invitedByName: string;
  invitedUserId: number;
  status: string;
  createdAt: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  streak: number;
  totalHours: number;
  groups: Group[];
  createdAt: string;
}