# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Studify API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List groups the current user belongs to
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new study group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupBody"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get a specific group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}/members:
    get:
      operationId: listGroupMembers
      tags: [groups]
      summary: List members of a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupMember"
  /groups/{groupId}/posts:
    get:
      operationId: listGroupPosts
      tags: [groups]
      summary: List posts (study sessions) in a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a study post in a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}/ranking:
    get:
      operationId: getGroupRanking
      tags: [groups]
      summary: Get weekly ranking for a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Weekly ranking
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RankingEntry"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        streak:
          type: integer
        totalHours:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - streak
        - totalHours
        - createdAt
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        adminId:
          type: integer
        adminName:
          type: string
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - adminId
        - adminName
        - memberCount
        - createdAt
    GroupMember:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        streak:
          type: integer
        totalHours:
          type: number
        isAdmin:
          type: boolean
      required:
        - userId
        - name
        - streak
        - totalHours
        - isAdmin
    CreateGroupBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    Post:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        subject:
          type: string
        hours:
          type: number
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - groupId
        - userId
        - userName
        - subject
        - hours
        - description
        - createdAt
    CreatePostBody:
      type: object
      properties:
        subject:
          type: string
        hours:
          type: number
        description:
          type: ["string", "null"]
      required:
        - subject
        - hours
    RankingEntry:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        weeklyHours:
          type: number
        streak:
          type: integer
        rank:
          type: integer
      required:
        - userId
        - name
        - weeklyHours
        - streak
        - rank
...