# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studify API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const RegisterBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  password: zod.string(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
    streak: zod.number(),
    totalHours: zod.number(),
    createdAt: zod.date(),
  }),
  message: zod.string(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  streak: zod.number(),
  totalHours: zod.number(),
  createdAt: zod.date(),
});

/**
 * @summary List groups the current user belongs to
 */
export const ListGroupsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  adminId: zod.number(),
  adminName: zod.string(),
  memberCount: zod.number(),
  createdAt: zod.date(),
});
export const ListGroupsResponse = zod.array(ListGroupsResponseItem);

/**
 * @summary Create a new study group
 */
export const CreateGroupBody = zod.object({
  name: zod.string(),
});

/**
 * @summary Get a specific group
 */
export const GetGroupParams = zod.object({
  groupId: zod.coerce.number(),
});

export const GetGroupResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  adminId: zod.number(),
  adminName: zod.string(),
  memberCount: zod.number(),
  createdAt: zod.date(),
});

/**
 * @summary List members of a group
 */
export const ListGroupMembersParams = zod.object({
  groupId: zod.coerce.number(),
});

export const ListGroupMembersResponseItem = zod.object({
  userId: zod.number(),
  name: zod.string(),
  streak: zod.number(),
  totalHours: zod.number(),
  isAdmin: zod.boolean(),
});
export const ListGroupMembersResponse = zod.array(ListGroupMembersResponseItem);

/**
 * @summary List posts (study sessions) in a group
 */
export const ListGroupPostsParams = zod.object({
  groupId: zod.coerce.number(),
});

export const ListGroupPostsResponseItem = zod.object({
  id: zod.number(),
  groupId: zod.number(),
  userId: zod.number(),
  userName: zod.string(),
  subject: zod.string(),
  hours: zod.number(),
  description: zod.string().nullable(),
  createdAt: zod.date(),
});
export const ListGroupPostsResponse = zod.array(ListGroupPostsResponseItem);

/**
 * @summary Create a study post in a group
 */
export const CreatePostParams = zod.object({
  groupId: zod.coerce.number(),
});

export const CreatePostBody = zod.object({
  subject: zod.string(),
  hours: zod.number(),
  description: zod.string().nullish(),
});

/**
 * @summary Get weekly ranking for a group
 */
export const GetGroupRankingParams = zod.object({
  groupId: zod.coerce.number(),
});

export const GetGroupRankingResponseItem = zod.object({
  userId: zod.number(),
  name: zod.string(),
  weeklyHours: zod.number(),
  streak: zod.number(),
  rank: zod.number(),
});
export const GetGroupRankingResponse = zod.array(GetGroupRankingResponseItem);

/**
 * @summary Send an invite to a user by email (admin only)
 */
export const SendInviteParams = zod.object({
  groupId: zod.coerce.number(),
});

export const SendInviteBody = zod.object({
  email: zod.string(),
});

/**
 * @summary List pending invites for the current user
 */
export const ListInvitesResponseItem = zod.object({
  id: zod.number(),
  groupId: zod.number(),
  groupName: zod.string(),
  invitedById: zod.number(),
  invitedByName: zod.string(),
  invitedUserId: zod.number(),
  status: zod.string(),
  createdAt: zod.date(),
});
export const ListInvitesResponse = zod.array(ListInvitesResponseItem);

/**
 * @summary Accept a group invite
 */
export const AcceptInviteParams = zod.object({
  inviteId: zod.coerce.number(),
});

export const AcceptInviteResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Reject a group invite
 */
export const RejectInviteParams = zod.object({
  inviteId: zod.coerce.number(),
});

export const RejectInviteResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get user profile
 */
export const GetUserProfileParams = zod.object({
  userId: zod.coerce.number(),
});

export const GetUserProfileResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  streak: zod.number(),
  totalHours: zod.number(),
  groups: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      adminId: zod.number(),
      adminName: zod.string(),
      memberCount: zod.number(),
      createdAt: zod.date(),
    }),
  ),
  createdAt: zod.date(),
});