# lib/api-zod/src/generated/types/createPostBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studify API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePostBody {
  subject: string;
  hours: number;
  /** @nullable */
  description?: string | null;
}