# lib/api-zod/src/generated/types/group.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studify API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Group {
  id: number;
  name: string;
  adminId: number;
  adminName: string;
  memberCount: number;
  createdAt: Date;
}