# lib/api-zod/src/generated/types/groupMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studify API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GroupMember {
  userId: number;
  name: string;
  streak: number;
  totalHours: number;
  isAdmin: boolean;
}