# lib/api-zod/src/generated/types/rankingEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studify API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RankingEntry {
  userId: number;
  name: string;
  weeklyHours: number;
  streak: number;
  rank: number;
}